pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "team-context-mcp"
version = "0.1.0"
description = "MCP server that injects shared team context into LLM sessions"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "sentence-transformers>=3.0.0",
    "sqlite-vec>=0.1.6",
    "click>=8.1.0",
    "gitpython>=3.1.40",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    # torch se instala vía install.sh con --index-url CPU-only.
    # No lo listamos aquí para no forzar la versión CUDA en pip install directo.
]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[project.scripts]
team-mcp = "team_context_mcp.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/team_context_mcp"]

[tool.hatch.metadata]
allow-direct-references = true
